requirements.txt
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic[email]==2.6.1
sqlalchemy==2.0.27
pymysql==1.1.0
bcrypt==4.1.2
PyJWT==2.10.1
email-validator==2.1.0.post1
razorpay==1.3.0 